dom/base/test/test_meta_viewport_no_width_and_valid_height.html

<!DOCTYPE HTML>
<html>
<head>
  <meta charset="utf-8">
  <title>valid height in meta viewport</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <meta name="viewport" content="height=240">
  <script src="viewport_helpers.js"></script>
</head>
<body>
  <p>height=240</p>
  <script type="application/javascript">
    "use strict";

    add_task(async function valid_height() {
      await SpecialPowers.pushPrefEnv(scaleRatio(1.0));

      let info = getViewportInfo(800, 480);
      is(info.width, 400, "width should be 400 which is the result of the " +
                           "display width * viewport height / display height");
      is(info.height, 240, "height should be 240");
    });
  </script>
</body>
</html>
